+++ changelog.md
@@ -11,6 +11,9 @@
 - Better Venus Atmosphere
 - Speed up fuel refining and water separation
 - Complete rewrite of the oxygen system
+- Change water separator gas tank size
+- Oxygen and hydrogen tank texture is now filling downward
+- Oxygen and hydrogen tank can now be filled even if one tank is full
 
 ## Fixes
 - Fix candle still burning in space